Add files via upload
Added files that were extracted from the game data files through Aury's NexusVault exporter. Included are the original texture maps, as well as the gltf file containing the original skeleton. WARNING: For whatever reason, said file also includes the skeleton and meshes from the Angel model, so if importing to blender, some trimming of the meshes and bones will be required.