qol improvements to rig
Added inverse kinematics and custom bone shapes to improve usability of rig.

Also improved simplicity of hands by introducing collective movement of fingers (in addition to separate).

The signia on top of Drusera's head was given a separate bone for possibility of in-depth animation. More natural parenting was also introduced.